component.js 파일추가
