+++ src/pages/dongseon/Login/Input.js
@@ -1,8 +1,13 @@
 import React from 'react';
 
-const Input = props => {
-  const { saveUserId, setSaveUserId, saveUserPassword, setSaveUserPassword } =
-    props;
+const Input = ({
+  saveUserId,
+  setSaveUserId,
+  saveUserPassword,
+  setSaveUserPassword,
+}) => {
+  // const { saveUserId, setSaveUserId, saveUserPassword, setSaveUserPassword } =
+  // props;
   return (
     <>
       <input
